+++ project/Dependencies.scala
@@ -4,7 +4,7 @@ object Dependencies {
 
   lazy val toplOrg = "co.topl"
 
-  lazy val bramblVersion = "2.0.0-beta0"
+  lazy val bramblVersion = "2.0.0-beta0+4-4f22b87e-SNAPSHOT"
   val bramblSdk = toplOrg %% "brambl-sdk" % bramblVersion 
 
   val monocleCore = "dev.optics" %% "monocle-core" % "3.2.0"