+++ src/main/scala/co/topl/brambl/cli/modules/GenusQueryModeModule.scala
@@ -5,6 +5,8 @@ import co.topl.brambl.cli.controllers.GenusQueryController
 import co.topl.brambl.dataApi.{GenusQueryAlgebra, RpcChannelResource}
 import co.topl.brambl.cli.BramblCliSubCmd
 import co.topl.brambl.cli.BramblCliParams
+import scopt.OParser
+import co.topl.brambl.cli.BramblCliParamsParserModule
 
 trait GenusQueryModeModule
     extends WalletStateAlgebraModule
@@ -14,7 +16,13 @@ trait GenusQueryModeModule
       validateParams: BramblCliParams
   ): IO[Either[String, String]] = validateParams.subcmd match {
     case BramblCliSubCmd.invalid =>
-      IO.pure(Left("A subcommand needs to be specified"))
+        IO.pure(
+          Left(
+            OParser.usage(
+              BramblCliParamsParserModule.genusQueryMode
+            ) + "\nA subcommand needs to be specified"
+          )
+        )
     case BramblCliSubCmd.utxobyaddress =>
       new GenusQueryController(
         walletStateAlgebra(